+++ packages/go_router/example/lib/others/extra_param.dart
@@ -2,44 +2,49 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// TODO(goderbauer): Refactor the examples to remove this ignore, https://github.com/flutter/flutter/issues/110210
-// ignore_for_file: avoid_dynamic_calls
-
-import 'dart:convert';
-
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
-final Map<String, dynamic> _families = const JsonDecoder().convert('''
-{
-  "f1": {
-    "name": "Doe",
-    "people": {
-      "p1": {
-        "name": "Jane",
-        "age": 23
-      },
-      "p2": {
-        "name": "John",
-        "age": 6
-      }
-    }
-  },
-  "f2": {
-    "name": "Wong",
-    "people": {
-      "p1": {
-        "name": "June",
-        "age": 51
-      },
-      "p2": {
-        "name": "Xin",
-        "age": 44
-      }
-    }
-  }
+/// Family data class.
+class Family {
+  /// Create a family.
+  const Family({required this.name, required this.people});
+
+  /// The last name of the family.
+  final String name;
+
+  /// The people in the family.
+  final Map<String, Person> people;
 }
-''');
+
+/// Person data class.
+class Person {
+  /// Creates a person.
+  const Person({required this.name, required this.age});
+
+  /// The first name of the person.
+  final String name;
+
+  /// The age of the person.
+  final int age;
+}
+
+const Map<String, Family> _families = <String, Family>{
+  'f1': Family(
+    name: 'Doe',
+    people: <String, Person>{
+      'p1': Person(name: 'Jane', age: 23),
+      'p2': Person(name: 'John', age: 6),
+    },
+  ),
+  'f2': Family(
+    name: 'Wong',
+    people: <String, Person>{
+      'p1': Person(name: 'June', age: 51),
+      'p2': Person(name: 'Xin', age: 44),
+    },
+  ),
+};
 
 void main() => runApp(App());
 
@@ -91,11 +96,11 @@ class HomeScreen extends StatelessWidget {
         appBar: AppBar(title: const Text(App.title)),
         body: ListView(
           children: <Widget>[
-            for (final String fid in _families.keys)
+            for (final MapEntry<String, Family> entry in _families.entries)
               ListTile(
-                title: Text(_families[fid]['name']),
-                onTap: () => context
-                    .goNamed('family', extra: <String, String>{'fid': fid}),
+                title: Text(entry.value.name),
+                onTap: () => context.goNamed('family',
+                    extra: <String, String>{'fid': entry.key}),
               )
           ],
         ),
@@ -112,15 +117,14 @@ class FamilyScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final Map<String, dynamic> people =
-        _families[fid]['people'] as Map<String, dynamic>;
+    final Map<String, Person> people = _families[fid]!.people;
     return Scaffold(
-      appBar: AppBar(title: Text(_families[fid]['name'])),
+      appBar: AppBar(title: Text(_families[fid]!.name)),
       body: ListView(
         children: <Widget>[
-          for (final dynamic p in people.values)
+          for (final Person p in people.values)
             ListTile(
-              title: Text(p['name']),
+              title: Text(p.name),
             ),
         ],
       ),